feat: security 권한 url 수정 및 post 리팩토링(#375)
