+++ backend/src/test/java/com/minibeit/post/web/PostByBusinessControllerTest.java
@@ -322,38 +322,6 @@ class PostByBusinessControllerTest extends MvcTest {
     }
 
     @Test
-    @DisplayName("게시물 실험이 있는 날짜 목록 조회(년,월 기준) 문서화")
-    public void getDoDateList() throws Exception {
-        Set<LocalDate> localDates = new HashSet<>();
-        localDates.add(LocalDate.of(2021, 9, 21));
-        localDates.add(LocalDate.of(2021, 9, 22));
-        localDates.add(LocalDate.of(2021, 9, 23));
-        PostResponse.DoDateList response = PostResponse.DoDateList.builder().doDateList(localDates).build();
-
-        given(postByBusinessService.getDoDateListByYearMonth(any(), any())).willReturn(response);
-
-        ResultActions results = mvc.perform(RestDocumentationRequestBuilders
-                .get("/api/post/{postId}/dates", 1)
-                .param("yearMonth", "2021-09"));
-
-        results.andExpect(status().isOk())
-                .andDo(print())
-                .andDo(document("post-doDate-list",
-                        pathParameters(
-                                parameterWithName("postId").description("게시물 식별자")
-                        ),
-                        requestParameters(
-                                parameterWithName("yearMonth").description("조회할 날짜")
-                        ),
-                        responseFields(
-                                fieldWithPath("status").type(JsonFieldType.NUMBER).description("상태 코드"),
-                                fieldWithPath("success").type(JsonFieldType.BOOLEAN).description("api 응답이 성공했다면 true"),
-                                fieldWithPath("data.doDateList[]").type(JsonFieldType.ARRAY).description("실험 있는 날짜")
-                        )
-                ));
-    }
-
-    @Test
     @DisplayName("게시물 삭제 문서화")
     public void deleteOne() throws Exception {
         ResultActions results = mvc.perform(RestDocumentationRequestBuilders.delete("/api/post/{postId}", 1));
